1-Sided PCB for Eagle
I've been making 1-sided PCBs for a long time. If you'd like to make a simple board for the Ratgdo, give this a try.
I used a 2N7000 and a RFP30N06LE transistor bought on Amazon.
No extra connectors, just a screw terminal, 3 10K resistors, some headers and a D1 Mini and you're all set.